consolidate OA and CNV dilution protocols to one folder
