--- protoBuilds/1610-CNV/README.json
@@ -1,1 +1,0 @@
-{"title": "CNV Dilution from CSV", "author": "Opentrons", "categories": {"Molecular Biology": ["DNA"]}, "description": "This protocol performs DNA dilutions on custom 96-well elution and dilution plates. The elution plate is mounted on a magnetic deck to separate beads. Transfer volumes are specified by a CSV file that can be uploaded on the custom protocol site. For reagent setup, see Additional Notes below.\nLinks:\nOA Dilution from CSV\nCNV Dilution from CSV\n\nYou will need:\n Thermo Fisher Scientific KingFisher 96 plate 200uL # 97002540\n USA Scientific TempPlate 96-well full skirt 0.1mL PCR plates # 1402-9800\n USA Scientific Channelmate 25ml reservoir # 1306-2590\n Opentrons P10 Single-Channel Electronic Pipette\n Opentrons P50 Single-Channel Electronic Pipette\n Opentrons 10ul Tips\n* Opentrons 300ul Tips", "robot": ["OT-2"], "modules": ["Opentrons Magnetic Module"], "reagents": ["MagMAX DNA Multi-Sample Ultra Kit # A25597", "Nuclease-free H2O"], "process": "\nUpload your CSV file, input your pipette mounts, and select the magnet setup for your elution plate.\nDownload your protocol.\nUpload your protocol into the OT App.\nSet up your deck according to the deck map.\nCalibrate your labware, tiprack and pipette using the OT App. For calibration tips, check out our support article.\nHit \"Run\".\nThe specified volumes of nuclease-free H2O is transferred to each well of the CNV dilution plate using the same tip. The pipette type is automatically selected based on transfer volume.\nThe specified volumes of OA plate contents are transferred to their corresponding wells on the CNV dilution plate using a new tip for each transfer. The pipette type is automatically selected based on the transfer volume.\n", "notes": "Reagent trough setup:\n* nuclease-free H2O: channel 1\n\nIf you have any questions about this protocol, please contact protocols@opentrons.com.", "internal": "ya76sE6P\n1610", "markdown": {"title": "CNV Dilution from CSV", "author": "[Opentrons](http://www.opentrons.com/)\n\n", "categories": "* Molecular Biology\n    * DNA\n\n", "description": "This protocol performs DNA dilutions on custom 96-well elution and dilution plates. The elution plate is mounted on a magnetic deck to separate beads. Transfer volumes are specified by a CSV file that can be uploaded on the custom protocol site. For reagent setup, see Additional Notes below.\n\nLinks:\n[OA Dilution from CSV](./1610-OA)\n[CNV Dilution from CSV](./1610-CNV)\n\n---\n\nYou will need:\n* [Thermo Fisher Scientific KingFisher 96 plate 200uL # 97002540](https://www.thermofisher.com/order/catalog/product/95040450)\n* [USA Scientific TempPlate 96-well full skirt 0.1mL PCR plates # 1402-9800](https://www.usascientific.com/full-skirted-96-well-PCR-plate.aspx)\n* [USA Scientific Channelmate 25ml reservoir # 1306-2590](https://www.usascientific.com/channelmate.aspx)\n* [Opentrons P10 Single-Channel Electronic Pipette](https://shop.opentrons.com/collections/ot-2-pipettes/products/single-channel-electronic-pipette)\n* [Opentrons P50 Single-Channel Electronic Pipette](https://shop.opentrons.com/collections/ot-2-pipettes/products/single-channel-electronic-pipette?variant=5984549077021)\n* [Opentrons 10ul Tips](https://shop.opentrons.com/collections/opentrons-tips/products/opentrons-10ul-tips)\n* [Opentrons 300ul Tips](https://shop.opentrons.com/collections/opentrons-tips/products/opentrons-300ul-tips)\n\n", "robot": "* [OT-2](https://opentrons.com/ot-2)\n\n", "modules": "* [Opentrons Magnetic Module](https://shop.opentrons.com/collections/hardware-modules/products/magdeck)\n\n", "reagents": "* [MagMAX DNA Multi-Sample Ultra Kit # A25597](https://www.thermofisher.com/order/catalog/product/A25597)\n* Nuclease-free H2O\n\n", "process": "1. Upload your CSV file, input your pipette mounts, and select the magnet setup for your elution plate.\n2. Download your protocol.\n3. Upload your protocol into the [OT App](https://opentrons.com/ot-app).\n4. Set up your deck according to the deck map.\n5. Calibrate your labware, tiprack and pipette using the OT App. For calibration tips, check out our [support article](https://support.opentrons.com/ot-2/getting-started-software-setup/deck-calibration).\n6. Hit \"Run\".\n7. The specified volumes of nuclease-free H2O is transferred to each well of the CNV dilution plate using the same tip. The pipette type is automatically selected based on transfer volume.\n8. The specified volumes of OA plate contents are transferred to their corresponding wells on the CNV dilution plate using a new tip for each transfer. The pipette type is automatically selected based on the transfer volume.\n\n", "notes": "Reagent trough setup:\n* nuclease-free H2O: channel 1\n\n---\n\nIf you have any questions about this protocol, please contact protocols@opentrons.com.\n\n", "internal": "ya76sE6P\n1610\n"}}
