--- protoBuilds/1610-CNV/cnv_dilution.ot2.py.json
@@ -1,1 +1,0 @@
-{"instruments": [{"name": "p50_single", "mount": "left"}, {"name": "p10_single", "mount": "right"}], "labware": [{"type": "USAScientific-96-PCR", "slot": "2", "name": "OA dilution plate", "share": false}, {"type": "USAScientific-96-PCR", "slot": "5", "name": "CNV dilution plate", "share": false}, {"type": "USAScientific-trough-1channel", "slot": "6", "name": "water trough", "share": false}, {"type": "tiprack-10ul", "slot": "4", "name": "tiprack-10ul", "share": false}, {"type": "tiprack-10ul", "slot": "7", "name": "tiprack-10ul", "share": false}, {"type": "tiprack-10ul", "slot": "8", "name": "tiprack-10ul", "share": false}, {"type": "opentrons-tiprack-300ul", "slot": "9", "name": "opentrons-tiprack-300ul", "share": false}, {"type": "opentrons-tiprack-300ul", "slot": "10", "name": "opentrons-tiprack-300ul", "share": false}, {"type": "opentrons-tiprack-300ul", "slot": "11", "name": "opentrons-tiprack-300ul", "share": false}], "modules": [], "parameters": [{"name": "csv_file", "annotation": {"type": "FileInput"}, "default": "011519 PGx - OA & CNV dilutions,,,,,,,,,\n,,,,OA,,,CNV,,\nWell,ID,Concentration,Well,ul DNA,ul H2O,Well,ul DNA,ul H2O,\nA1,NTC,6.1,A1,30.0,0.0,A1,16.3,3.7,\nA2,EXT_CTRL_2,447.4,A2,3.4,26.6,A2,2.0,18.0,\nA3,MD11636,11.8,A3,30.0,0.0,A3,8.5,11.5,\nA4,MD11637,9.0,A4,30.0,0.0,A4,11.1,8.9,\nA5,MD11638,15.3,A5,30.0,0.0,A5,6.5,13.5,\nA6,MD11639,16.9,A6,30.0,0.0,A6,5.9,14.1,\nA7,MD11640,58.8,A7,25.5,4.5,A7,2.0,18.0,\nA8,MD11641,26.5,A8,30.0,0.0,A8,3.8,16.2,\nA9,MD11642,37.6,A9,30.0,0.0,A9,2.7,17.3,\nA10,MD11643,15.9,A10,30.0,0.0,A10,6.3,13.7,\nA11,MD11644,11.5,A11,30.0,0.0,A11,8.7,11.3,\nA12,PHI17444,69.3,A12,21.6,8.4,A12,2.0,18.0,\nB1,PHI17445,11.8,B1,30.0,0.0,B1,8.5,11.5,\n"}, {"name": "P10_mount", "annotation": {"type": "StringSelection", "options": [{"value": "right", "text": "right"}, {"value": "left", "text": "left"}]}, "default": "right"}, {"name": "P50_mount", "annotation": {"type": "StringSelection", "options": [{"value": "left", "text": "left"}, {"value": "right", "text": "right"}]}, "default": "left"}]}
