--- protoBuilds/1610-CNV/metadata.json
@@ -1,1 +1,0 @@
-{"slug": "1610-CNV", "path": "protocols/1610-CNV", "flags": {"feature": false, "skip-tests": false, "embedded-app": false}, "files": {"description": ["README.md"], "OT 1 protocol": [], "OT 2 protocol": ["cnv_dilution.ot2.py"]}, "status": "ok"}
