--- protoBuilds/1610-OA/metadata.json
@@ -1,1 +1,0 @@
-{"slug": "1610-OA", "path": "protocols/1610-OA", "flags": {"feature": false, "skip-tests": false, "embedded-app": false}, "files": {"description": ["README.md"], "OT 1 protocol": [], "OT 2 protocol": ["oa_dilution.ot2.py"]}, "status": "ok"}
